+++ testapps/GHIssues/src/main/java/ghissues/gh820/BaseService.java
@@ -1,0 +1,7 @@
+package ghissues.gh820;
+
+import java.util.List;
+
+public interface BaseService<T> {
+    List<T> produceList();
+}