+++ testapps/GHIssues/src/main/java/ghissues/gh820/Gh820StringService.java
@@ -1,8 +1,0 @@
-package ghissues.gh820;
-
-public class Gh820StringService implements Gh820Service<String> {
-    @Override
-    public String get() {
-        return "abc";
-    }
-}