+++ testapps/GHIssues/src/main/java/ghissues/gh820/IntegerService.java
@@ -1,0 +1,12 @@
+package ghissues.gh820;
+
+import org.osgl.util.C;
+
+import java.util.List;
+
+public class IntegerService implements BaseService<Integer> {
+    @Override
+    public List<Integer> produceList() {
+        return C.list(1, 2, 3);
+    }
+}